Proposed fix for issue #36
This is my solution for issue #36 (decimal point on numpad not working).
The keycode for the numpad decimal point was translated to the keycode for "delete". I removed the translation because in IE, FF and chrome the key on the numpad gives the keycode for "decimal point" when the numlock is on, and delete when it is off.

Further down, I added an extra if-condition to translate the decimal point key to whatever character has been set as the decimal separator.