Always reset cursor position
If all the context of a input field are selected, always reset the cursor position.
e.g. When the cursor is between the second and third decimal place, and the user presses Ctrl+A, and then types in a number. In that case the cursor should not stay between the second and third decimal place: it should be moved to just before the decimal point.